Code for Detection and Classification of cancers
aims at developing ML models that give an accurate result and helps in detecting a tumor in the brain, breast and skin from
the input image (MRI, Histopathological, and skin patches) using CNN. The model output is further available on GUI,
making it user-friendly and easy for diagnosis. The result obtained can be used by the medical fraternity for the early
detection and diagnosis of cancer.